app/requirements.txt
requests
openai
boto3
google-generativeai>=0.8.3
Pillow